requirements.txt
chainlit
notebook
ipywidgets
tqdm
openai
python-dotenv

-e .